Move actors to a separate package
