+++ src/main/java/com/codecool/quest/logic/Cell.java
@@ -1,5 +1,7 @@
 package com.codecool.quest.logic;
 
+import com.codecool.quest.logic.actors.Actor;
+
 public class Cell implements Drawable {
     private CellType type = CellType.EMPTY;
     private Actor actor;
