+++ src/main/java/com/codecool/quest/logic/MapLoader.java
@@ -1,5 +1,8 @@
 package com.codecool.quest.logic;
 
+import com.codecool.quest.logic.actors.Player;
+import com.codecool.quest.logic.actors.Skeleton;
+
 import java.io.InputStream;
 import java.util.Scanner;
