+++ src/main/java/com/codecool/quest/logic/actors/Actor.java
@@ -1,4 +1,7 @@
-package com.codecool.quest.logic;
+package com.codecool.quest.logic.actors;
+
+import com.codecool.quest.logic.Cell;
+import com.codecool.quest.logic.Drawable;
 
 public abstract class Actor implements Drawable {
     private Cell cell;
