+++ src/main/java/com/codecool/quest/logic/actors/Skeleton.java
@@ -1,4 +1,7 @@
-package com.codecool.quest.logic;
+package com.codecool.quest.logic.actors;
+
+import com.codecool.quest.logic.Cell;
+import com.codecool.quest.logic.actors.Actor;
 
 public class Skeleton extends Actor {
     public Skeleton(Cell cell) {
